test: Separate test/integration/integration.h into 3 different classes (#12794)
