+++ test/integration/fake_upstream.h
@@ -7,14 +7,11 @@
 
 #include "envoy/api/api.h"
 #include "envoy/config/core/v3/base.pb.h"
-#include "envoy/event/timer.h"
 #include "envoy/grpc/status.h"
 #include "envoy/http/codec.h"
 #include "envoy/network/connection.h"
 #include "envoy/network/connection_handler.h"
 #include "envoy/network/filter.h"
-#include "envoy/server/configuration.h"
-#include "envoy/server/listener_manager.h"
 #include "envoy/stats/scope.h"
 
 #include "common/buffer/buffer_impl.h"
@@ -26,7 +23,6 @@
 #include "common/common/thread.h"
 #include "common/grpc/codec.h"
 #include "common/grpc/common.h"
-#include "common/http/exception.h"
 #include "common/http/http1/codec_impl.h"
 #include "common/http/http2/codec_impl.h"
 #include "common/network/connection_balancer_impl.h"
@@ -37,7 +33,6 @@
 
 #include "server/active_raw_udp_listener_config.h"
 
-#include "test/test_common/printers.h"
 #include "test/test_common/test_time_system.h"
 #include "test/test_common/utility.h"
