+++ test/integration/server.cc
@@ -18,8 +18,8 @@
 #include "server/process_context_impl.h"
 
 #include "test/common/runtime/utility.h"
-#include "test/integration/integration.h"
 #include "test/integration/utility.h"
+#include "test/mocks/common.h"
 #include "test/mocks/runtime/mocks.h"
 #include "test/test_common/environment.h"
